Rename abs_encoder.py to abc
